Update PlotFigureManager to load the .ui file once.
In addition this changeset also removes the BaseQtPlotWindow class
as it basically did nothing.
The matplotlib figure canvas and toolbar are now imported based on the
current version of Qt rather than being hardcoded to Qt4.
